+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.99.0
+
 - Truncate tool call outputs automatically to avoid hit context size limit. #284
   - Save output to eca cache folde and tell LLM where the full content is.
   - Add `toolCall outputTruncation` config to customize size or lines.